+++ cv_pipeliner/__init__.py
@@ -20,6 +20,7 @@ from cv_pipeliner.inference_models.detection.object_detection_api import (
     ObjectDetectionAPI_pb_ModelSpec
 )
 from cv_pipeliner.inference_models.detection.pytorch import PytorchDetection_ModelSpec
+from cv_pipeliner.inference_models.detection.yolov5 import YOLOv5_ModelSpec
 from cv_pipeliner.inference_models.classification.tensorflow import (
     TensorFlow_ClassificationModelSpec, TensorFlow_ClassificationModelSpec_TFServing
 )
@@ -27,3 +28,9 @@ from cv_pipeliner.inference_models.pipeline import PipelineModelSpec, PipelineMo
 from cv_pipeliner.inferencers.detection import DetectionInferencer
 from cv_pipeliner.inferencers.classification import ClassificationInferencer
 from cv_pipeliner.inferencers.pipeline import PipelineInferencer
+
+
+from cv_pipeliner.data_converters.coco import COCODataConverter
+from cv_pipeliner.data_converters.brickit import BrickitDataConverter
+from cv_pipeliner.data_converters.json import JSONDataConverter
+from cv_pipeliner.data_converters.supervisely import SuperviselyDataConverter